+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 ## 0.4.0 (In Development)
   - `VisionARManager` allows to change ARLane's length
+  - `boundingBox` property on `MBVDetection` now stores normalized relative coordinates